Add comment to morph.py and add folder structure to this code
